requirements.txt
numpy
opencv-python-headless
scikit-image
scipy
pillow
pandas
networkx
matplotlib